% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_annotation.R
\name{AnnoPeak}
\alias{AnnoPeak}
\title{Conduct Peak Annotation.}
\usage{
AnnoPeak(
  peak.df,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  seq.style = c("UCSC", "NCBI", "Ensembl", "None"),
  gtf.file = NULL,
  up.dist = 3000,
  down.dist = 3000,
  ...
)
}
\arguments{
\item{peak.df}{Dataframe contains all consensus peaks.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{seq.style}{The style of sequence, chosen from UCSC, NCBI, Ensembl, None. This should be compatible with the genome and gtf file you used
to generate count matrix and peak files. Default: "UCSC".}

\item{gtf.file}{GTF file used to create TxDb object. Useful when specie you used is not available in \code{species}. Default: NULL.}

\item{up.dist}{The upstream distance from the TSS. Default: 3000bp.}

\item{down.dist}{The downstream distance from the TSS. Default: 3000bp.}

\item{...}{Parameters for \code{\link{annotatePeak}}.}
}
\value{
List contains peak annotation dataframe and plot.
}
\description{
Conduct Peak Annotation.
}
\examples{
library(DEbPeak)
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
}
