% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{ContinuousColorMap}
\alias{ContinuousColorMap}
\title{Create Continuous Color Map.}
\usage{
ContinuousColorMap(
  n = 50,
  direction = 1,
  maps = NULL,
  low = "white",
  high = "red",
  mid = NULL,
  ...
)
}
\arguments{
\item{n}{The number of colors (≥ 1) to be in the palette.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{maps}{Existing color maps, chosen from "magma","inferno","plasma","viridis","cividis","rocket","mako","turbo".
Default: NULL (use \code{low}, \code{high}, \code{mid} to create).}

\item{low}{The low color. Default: white.}

\item{high}{The high color. Default: red.}

\item{mid}{The mid color. Default: NULL.}

\item{...}{Parameters for \code{\link{viridis}}.}
}
\value{
A vector of colors.
}
\description{
Create Continuous Color Map.
}
\examples{
# use existing color map
ContinuousColorMap(maps = "viridis")
# create color map with low, high, mid.
ContinuousColorMap(maps = NULL, low = "magenta", high = "yellow", mid = "black")
}
