% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_qc.R
\name{CountQC}
\alias{CountQC}
\title{Count QC plot.}
\usage{
CountQC(
  deobj,
  group.key = NULL,
  type = c("saturation", "cpm"),
  min.count = 0,
  ndepth = 10,
  cat.colors = NULL,
  legend.col = NULL,
  legend.inset = -0.2,
  ...
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{group.key}{Sample group information. When set NULL, select first column of metadata. Default: NULL.}

\item{type}{QC plot type, chosen from saturation and cpm. Default: saturation.}

\item{min.count}{A feature is considered to be detected if the corresponding number of read counts is > \code{min.count}. By default, \code{min.count} = 0. . This parameter is used by type "saturation".}

\item{ndepth}{Number of different sequencing depths to be simulated and plotted apart from the real depth. Default: 10. This parameter is only used by type "saturation".}

\item{cat.colors}{Color used for different CPM Threshold group. Default: NULL.}

\item{legend.col}{Column number of legend. Default: NULL(length of CPM Threshold group).}

\item{legend.inset}{Inset distance(s) from the margins as a fraction of the plot region when legend is placed by keyword. Default: -0.2.}

\item{...}{Parameters for \code{\link{dat}}.}
}
\value{
Count QC plot
}
\description{
Count QC plot.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
CountQC(deobj = dds, group.key = "condition", type = "cpm")
CountQC(deobj = dds, group.key = "condition", type = "saturation")
}
