% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_CA.R
\name{DEbCA}
\alias{DEbCA}
\title{Integrate Differential Expression Results and Peak Annotation Results of ChIP-seq and ATAC-seq.}
\usage{
DEbCA(
  de.res,
  chip.peak.res,
  atac.peak.res,
  peak.anno.key = c("Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream",
    "Distal Intergenic", "All"),
  signif = "padj",
  signif.threashold = 0.05,
  l2fc.threashold = 1,
  label.key = NULL,
  merge.key = c("geneId", "ENSEMBL", "SYMBOL")
)
}
\arguments{
\item{de.res}{Data frame contains all genes of differential expression analysis.}

\item{chip.peak.res}{Dataframe contains peak annotation results of ChIP-seq data.}

\item{atac.peak.res}{Dataframe contains peak annotation results of ATAC-seq data.}

\item{peak.anno.key}{Peak location, chosen from "Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream", "Distal Intergenic","All". Default: "Promoter".}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threashold}{Significance threashold to get differentially expressed genes. Default: 0.05.}

\item{l2fc.threashold}{Log2 fold change threashold to get differentially expressed genes. Default: 1.}

\item{label.key}{Which column to use as label. Default: NULL (use rownames of \code{deres}).}

\item{merge.key}{The columns used for merging, chosen from geneId, ENSEMBL, SYMBOL. Default: geneId.}
}
\value{
Dataframe contains integrated results. The results will contain eleven categories: UP, DOWN, ChIP, ATAC, UPbChIP, DOWNbChIP
UPbATAC, DOWNbATAC, ChIPbATAC, UPbPeak, DOWNbPeak.
}
\description{
Integrate Differential Expression Results and Peak Annotation Results of ChIP-seq and ATAC-seq.
}
\examples{
library(DEbPeak)
library(DESeq2)
# ChIP-Seq data
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# RNA-Seq data
count.file <- system.file("extdata", "debchip_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "debchip_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
# create DESeqDataSet object
dds <- DESeq2::DESeqDataSetFromMatrix(
  countData = count.matrix, colData = meta.info,
  design = ~condition
)
# set control level
dds$condition <- relevel(dds$condition, ref = "NF")
# conduct differential expressed genes analysis
dds <- DESeq(dds)
# extract results
dds.results <- results(dds, contrast = c("condition", "RX", "NF"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
# Integrated with RNA-Seq
debchip.res <- DEbPeak(
  de.res = dds.results.ordered, peak.res = peak.anno[["df"]],
  peak.anno.key = "Promoter", merge.key = "SYMBOL"
)
# ATAC-seq data
atac.peak.file <- system.file("extdata", "debatac_peaks.bed", package = "DEbPeak")
atac.peak.df <- GetConsensusPeak(peak.file = atac.peak.file)
atac.peak.profile <- PeakProfile(atac.peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
atac.peak.anno <- AnnoPeak(
  peak.df = atac.peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
debpeak.res <- DEbCA(
  de.res = dds.results.ordered, chip.peak.res = peak.anno[["df"]],
  atac.peak.res = atac.peak.anno[["df"]],
  peak.anno.key = "Promoter", merge.key = "SYMBOL"
)
}
