% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_DE.R
\name{DEbDE}
\alias{DEbDE}
\title{Integrate Two Differential Expression Results.}
\usage{
DEbDE(
  de1.res,
  de2.res,
  de1.signif = "padj",
  de1.signif.threshold = 0.05,
  de1.l2fc.threshold = 1,
  de2.signif = "padj",
  de2.signif.threshold = 0.05,
  de2.l2fc.threshold = 1
)
}
\arguments{
\item{de1.res}{DE1 dataframe contains all genes of differential expression analysis.}

\item{de2.res}{DE2 dataframe contains all genes of differential expression analysis.}

\item{de1.signif}{Significance criterion for DE1. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{de1.signif.threshold}{Significance threshold for DE1 to get differentially expressed genes. Default: 0.05.}

\item{de1.l2fc.threshold}{Log2 fold change threshold for DE1 to get differentially expressed genes. Default: 1.}

\item{de2.signif}{Significance criterion for DE2. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{de2.signif.threshold}{Significance threshold for DE2 to get differentially expressed genes. Default: 0.05.}

\item{de2.l2fc.threshold}{Log2 fold change threshold for DE2 to get differentially expressed genes. Default: 1.}
}
\value{
Dataframe contains integration results, the 'Type' column contains "Down_Up", "Up_Up", "Down_Down", "Up_Down",
"DE1_Up", "DE1_Down", "DE2_Up", "DE2_Down", "Not_Not".
}
\description{
Integrate Two Differential Expression Results.
}
\examples{
library(DEbPeak)
rna.diff.file <- system.file("extdata", "RA_RNA_diff.txt", package = "DEbPeak")
de1.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
de2.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
# use same file as example
de.de <- DEbDE(de1.res = de1.res, de2.res = de2.res, de1.l2fc.threshold = 0.5, de2.l2fc.threshold = 1)
}
