% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_DE.R
\name{DEbDEFE}
\alias{DEbDEFE}
\title{GO Enrichment on Two Differential Expression Integration Results.}
\usage{
DEbDEFE(
  de.de,
  de.fe.key,
  out.folder = NULL,
  gene.type = c("ENSEMBL", "ENTREZID", "SYMBOL"),
  go.type = c("ALL", "BP", "MF", "CC"),
  enrich.pvalue = 0.05,
  enrich.qvalue = 0.05,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  padj.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  show.term = 15,
  str.width = 30,
  plot.resolution = 300,
  plot.width = 7,
  plot.height = 9,
  save = TRUE
)
}
\arguments{
\item{de.de}{Dataframe contains integrated results of two differential expression analysis of RNA-seq.}

\item{de.fe.key}{The key type of integrated results ("Type" column of \code{de.de}) to perform functional enrichment.}

\item{out.folder}{Folder to save enrichment results. Default: wording directory.}

\item{gene.type}{Gene name type of \code{DE1_Gene} column. Chosen from ENSEMBL, ENTREZID,SYMBOL. Default: ENSEMBL.}

\item{go.type}{GO enrichment type, chosen from ALL, BP, MF, CC. Default: ALL.}

\item{enrich.pvalue}{Cutoff value of pvalue. Default: 0.05.}

\item{enrich.qvalue}{Cutoff value of qvalue. Default: 0.05.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{padj.method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: BH.}

\item{show.term}{Number of enrichment term to show. Default: 15.}

\item{str.width}{Length of enrichment term in plot. Default: 30.}

\item{plot.resolution}{Resolution of plot. Default: 300.}

\item{plot.width}{The width of plot. Default: 7.}

\item{plot.height}{The height of plot. Default: 9.}

\item{save}{Logical value, whether to save all results. Default: TRUE.}
}
\value{
If \code{save} is TRUE, return NULL (all results are in \code{out.folder}), else retutn result dataframe.
}
\description{
GO Enrichment on Two Differential Expression Integration Results.
}
\examples{
library(DEbPeak)
rna.diff.file <- system.file("extdata", "RA_RNA_diff.txt", package = "DEbPeak")
de1.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
de2.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
# use same file as example
de.de <- DEbDE(de1.res = de1.res, de2.res = de2.res, de1.l2fc.threshold = 0.5, de2.l2fc.threshold = 1)
de.de.fe <- DEbDEFE(de.de = de.de, de.fe.key = "Down_Down", gene.type = "SYMBOL", go.type = "BP", species = "Mouse", save = F)
}
