% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_Peak.R
\name{DEbPeakFE}
\alias{DEbPeakFE}
\title{GO Enrichment on Integrated Results.}
\usage{
DEbPeakFE(
  de.peak,
  peak.fe.key,
  out.folder = NULL,
  go.type = c("ALL", "BP", "MF", "CC"),
  enrich.pvalue = 0.05,
  enrich.qvalue = 0.05,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  padj.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  show.term = 15,
  str.width = 30,
  plot.resolution = 300,
  plot.width = 7,
  plot.height = 9,
  save = TRUE
)
}
\arguments{
\item{de.peak}{Dataframe contains integrated results.}

\item{peak.fe.key}{The key type of integrated results ("Type" column of \code{de.peak}) to perform functional enrichment.}

\item{out.folder}{Folder to save enrichment results. Default: wording directory.}

\item{go.type}{GO enrichment type, chosen from ALL, BP, MF, CC. Default: ALL.}

\item{enrich.pvalue}{Cutoff value of pvalue. Default: 0.05.}

\item{enrich.qvalue}{Cutoff value of qvalue. Default: 0.05.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{padj.method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: BH.}

\item{show.term}{Number of enrichment term to show. Default: 15.}

\item{str.width}{Length of enrichment term in plot. Default: 30.}

\item{plot.resolution}{Resolution of plot. Default: 300.}

\item{plot.width}{The width of plot. Default: 7.}

\item{plot.height}{The height of plot. Default: 9.}

\item{save}{Logical value, whether to save all results. Default: TRUE.}
}
\value{
If \code{save} is TRUE, return NULL (all results are in \code{out.folder}), else retutn result dataframe.
}
\description{
GO Enrichment on Integrated Results.
}
\examples{
library(DEbPeak)
library(DESeq2)
# ChIP-Seq data
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# RNA-Seq data
count.file <- system.file("extdata", "debchip_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "debchip_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
# create DESeqDataSet object
dds <- DESeq2::DESeqDataSetFromMatrix(
  countData = count.matrix, colData = meta.info,
  design = ~condition
)
# set control level
dds$condition <- relevel(dds$condition, ref = "NF")
# conduct differential expressed genes analysis
dds <- DESeq(dds)
# extract results
dds.results <- results(dds, contrast = c("condition", "RX", "NF"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
# Integrated with RNA-Seq
debchip.res <- DEbPeak(
  de.res = dds.results.ordered, peak.res = peak.anno[["df"]],
  peak.anno.key = "Promoter", merge.key = "SYMBOL"
)
# functional enrichment on UPbPeak genes
upbpeak.fe.results <- DEbPeakFE(
  de.peak = debchip.res, peak.fe.key = "UPbPeak",
  species = "Mouse", save = FALSE
)
# functional enrichment on DOWNbPeak genes
downbpeak.fe.results <- DEbPeakFE(
  de.peak = debchip.res, peak.fe.key = "DOWNbPeak",
  species = "Mouse", save = FALSE
)
}
