% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_visualization.R
\name{DEHeatmap}
\alias{DEHeatmap}
\title{Heatmap for Differentially Expressed Genes.}
\usage{
DEHeatmap(
  deobj,
  deres,
  group.key = NULL,
  ref.group = NULL,
  group.color = c("blue", "red"),
  gene.df = NULL,
  label.key = NULL,
  signif = "padj",
  signif.threashold = 0.05,
  l2fc.threashold = 1,
  exp.range = c(-2, 2),
  exp.color = c("green", "black", "red"),
  heatmap.height = 20,
  heatmap.width = 20,
  col.gap = 2,
  legend.height = 5,
  link.height = 4
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{deres}{Data frame contains all genes.}

\item{group.key}{Sample group information. When set NULL, select first column of metadata. Default: NULL.}

\item{ref.group}{Reference group name. When set NULL, select first element of groups. Default: NULL.}

\item{group.color}{Color for different sample group. Default: blue for \code{ref.group}, and red for the other group.}

\item{gene.df}{Gene data frame, at least contains Gene column. Default: NULL. When provided,
the second column should not be in \code{deres}.}

\item{label.key}{Which column to use as label. Default: NULL. When set NULL, use rownames of \code{deres} or Gene column of \code{gene.df}.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threashold}{Significance threashold to get differentially expressed genes. Default: 0.05.}

\item{l2fc.threashold}{Log2 fold change threashold to get differentially expressed genes. Default: 1.}

\item{exp.range}{Z-score range to plot. Default: c(-2,2).}

\item{exp.color}{Color map used for heatmap. Default: c("green","black","red").}

\item{heatmap.height}{The height of whole heatmap. Default: 20cm.}

\item{heatmap.width}{The width of whole heatmap. Default: 20cm.}

\item{col.gap}{Gap between column slices. Default: 2mm.}

\item{legend.height}{The height of legend. Default: 5cm.}

\item{link.height}{The height of the segments. Default: 4mm.}
}
\value{
A \code{\link{Heatmap-class}}.
}
\description{
Heatmap for Differentially Expressed Genes.
}
\examples{
library(DESeq2)
library(DEbPeak)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
dds$condition <- relevel(dds$condition, ref = "WT")
dds <- DESeq(dds)
dds.results <- results(dds, contrast = c("condition", "KO", "WT"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
DEHeatmap(deobj = dds, deres = dds.results.ordered, group.key = "condition", ref.group = "WT", signif = "pvalue", l2fc.threashold = 0.3)
}
