% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_visualization.R
\name{DiffPeakPie}
\alias{DiffPeakPie}
\title{Stat Genomic Regions of Differential Peaks with Pie Plot.}
\usage{
DiffPeakPie(
  deres,
  signif = "padj",
  signif.threshold = 0.05,
  l2fc.threshold = 1
)
}
\arguments{
\item{deres}{Data frame contains all peaks.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threshold}{Significance threshold to get differentially accessible/binding peaks. Default: 0.05.}

\item{l2fc.threshold}{Log2 fold change threshold to get differentially accessible/binding peaks. Default: 1.}
}
\value{
List contains ggplot2 objects.
}
\description{
Stat Genomic Regions of Differential Peaks with Pie Plot.
}
