% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palette.R
\name{DiscreteColorPalette}
\alias{DiscreteColorPalette}
\title{Create Discrete Color Palette.}
\usage{
DiscreteColorPalette(
  palette = c("dark2_8", "accent_8", "set2_8", "set1_9", "vega_10_scanpy", "set3_12",
    "vega_20_scanpy", "stepped_24_seurat", "alphabet_26_seurat", "alphabet2_26_seurat",
    "zeileis_28_scanpy", "glasbey_32_seurat", "polychrome_36_seurat", "cbf_40_dittoseq",
    "godsnot_102_scanpy")
)
}
\arguments{
\item{palette}{Palette used, chosen from "dark2_8", "accent_8", "set2_8", "set1_9", "vega_10_scanpy",
"set3_12", "vega_20_scanpy", "stepped_24_seurat", "alphabet_26_seurat", "alphabet2_26_seurat",
"zeileis_28_scanpy", "glasbey_32_seurat", "polychrome_36_seurat", "cbf_40_dittoseq", "godsnot_102_scanpy".
The last number indicates the length of this palette. Default: "dark2_8".}
}
\value{
A vector of colors.
}
\description{
Create Discrete Color Palette.
}
\examples{
library(DEbPeak)
DiscreteColorPalette(palette = "dark2_8")
}
