% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{DownloadGEOSupp}
\alias{DownloadGEOSupp}
\title{Download GEO Supplementary Files.}
\usage{
DownloadGEOSupp(acce, supp.idx = 1, out.folder = NULL)
}
\arguments{
\item{acce}{GEO accession number.}

\item{supp.idx}{The index of supplementary files to download. If NULL, download all supplementary files. Default: 1.}

\item{out.folder}{The output folder used to save downloaded results.}
}
\value{
A vector contains paths.
}
\description{
Download GEO Supplementary Files.
}
\examples{
# DownloadGEOSupp(acce = "GSE149836", supp.idx = NULL, out.folder = "/path/to/suppfiles")
}
