% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_enrichment.R
\name{EnrichPlot}
\alias{EnrichPlot}
\title{Create Plot for Functional Enrichment Analysis.}
\usage{
EnrichPlot(
  fe.res,
  signif = "p.adjust",
  show.num = 10,
  grad.col.vec = c("grey", "red"),
  fill.color = "red",
  str.width = 30,
  plot.type = c("bar", "dot"),
  plot.resolution = 300,
  plot.width = 7,
  plot.height = 9,
  save = FALSE
)
}
\arguments{
\item{fe.res}{Functional enrichment analysis results to plot, usually selected from raw results.}

\item{signif}{Significance criterion of the results, choose from p.adjust, pvalue, qvalue,
NULL (do not show significance information). Default: p.adjust.}

\item{show.num}{The number of terms to show, select in descending order of the FoldEnrichment. Default: 10.}

\item{grad.col.vec}{Two color gradient vector (low, high) for \code{signif}.
Used when \code{signif} is not NULL. Default: c("grey", "red").}

\item{fill.color}{Color used to color the geom when \code{signif} is NULL. Default: "red".}

\item{str.width}{Length of enrichment term in plot. If NULL, use raw length. Default: 30.}

\item{plot.type}{The type of the plot, choose from bar and dot. Default: bar.}

\item{plot.resolution}{Resolution of plot. Default: 300.}

\item{plot.width}{The width of plot. Default: 7.}

\item{plot.height}{The height of plot. Default: 9.}

\item{save}{Logical value, whether to save all results. Default: FALSE.}
}
\value{
If \code{save} is TRUE, return NULL, else return a ggplot2 object.
}
\description{
Create Plot for Functional Enrichment Analysis.
}
\examples{
# show p.adjust
# EnrichPlot(fe.res = fe.res, grad.col.vec=c("grey", "red"))
# EnrichPlot(fe.res = fe.res, grad.col.vec=c("grey", "red"), plot.type = "dot")
# do not show signif
# EnrichPlot(fe.res = fe.res, signif = NULL, fill.color = "red")
# EnrichPlot(fe.res = fe.res, signif = NULL, fill.color = "red", plot.type = "dot")
}
