% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{ExportPCGenes}
\alias{ExportPCGenes}
\title{Export selected PC genes.}
\usage{
ExportPCGenes(
  pca,
  data.type = c("RNA", "ChIP", "ATAC"),
  peak.anno.key = c("Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream",
    "Distal Intergenic", "All"),
  pc = 1:5,
  gene.num = 10
)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{data.type}{Input data type, choose from RNA, ChIP, ATAC. Default: RNA.}

\item{peak.anno.key}{Peak location, chosen from "Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream", "Distal Intergenic","All". Default: "Promoter".}

\item{pc}{Specify PC to export genes. Default: 1:5.}

\item{gene.num}{Gene number to export for every PC. Default: 10.}
}
\value{
Dataframe contains PC, Gene and Type (Positive or Negative).
}
\description{
Export selected PC genes.
}
\examples{
library(DESeq2)
library(DEbPeak)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
ExportPCGenes(pca = pca_res, pc = 1:2, gene.num = 200)
}
