% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_visualization.R
\name{ExtractDEG}
\alias{ExtractDEG}
\title{Extract Differentially Expressed Genes.}
\usage{
ExtractDEG(
  deres,
  signif = "padj",
  signif.threashold = 0.05,
  l2fc.threashold = 1
)
}
\arguments{
\item{deres}{Data frame contains all genes.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threashold}{Significance threashold to get differentially expressed genes. Default: 0.05.}

\item{l2fc.threashold}{Log2 fold change threashold to get differentially expressed genes. Default: 1.}
}
\value{
A data frame.
}
\description{
Extract Differentially Expressed Genes.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
dds$condition <- relevel(dds$condition, ref = "WT")
dds <- DESeq(dds)
dds.results <- results(dds, contrast = c("condition", "KO", "WT"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
dds.degs <- ExtractDEG(
  deres = dds.results.ordered, signif = "padj", signif.threashold = 0.05,
  l2fc.threashold = 1
)
}
