% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{ExtractGEOExp}
\alias{ExtractGEOExp}
\title{Extract Raw Count Matrix.}
\usage{
ExtractGEOExp(pf.obj, acce, supp.idx = 1, down.supp = FALSE, timeout = 3600)
}
\arguments{
\item{pf.obj}{GEO object of platform.}

\item{acce}{GEO accession number.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}

\item{down.supp}{Logical value, whether to download supplementary files to create count matrix. If TRUE, always
download supplementary files. If FALSE, use \code{ExpressionSet} (If contains non-integer or emoty,
download supplementary files automatically). Default: FALSE.}

\item{timeout}{Timeout for \code{\link{download.file}}. Default: 3600.}
}
\value{
A dataframe.
}
\description{
Extract Raw Count Matrix.
}
\examples{
# pf.obj = GEOobj(acce = "GSE147507", platform = "GPL28369")
# count.mat = ExtractGEOExp(pf.obj, acce = "GSE147507")
}
