% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{ExtractGEOExpSupp}
\alias{ExtractGEOExpSupp}
\title{Extract Raw Count Matrix from Supplementary Files.}
\usage{
ExtractGEOExpSupp(acce, supp.idx = 1)
}
\arguments{
\item{acce}{GEO accession number.}

\item{supp.idx}{The index of supplementary files to download. Default: 1.}
}
\value{
A dataframe.
}
\description{
Extract Raw Count Matrix from Supplementary Files.
}
\examples{
# count.mat = ExtractGEOExpSupp(acce = "GSE149838")
# count.mat = ExtractGEOExpSupp(acce = "GSE147507")
# count.mat = ExtractGEOExpSupp(acce = "GSE147507", supp.idx = 2)
# count.mat = ExtractGEOExpSupp(acce = "GSE122774")
}
