% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{ExtractGEOInfo}
\alias{ExtractGEOInfo}
\title{Extract GEO Study Information.}
\usage{
ExtractGEOInfo(pf.obj, sample.wise = FALSE)
}
\arguments{
\item{pf.obj}{GEO object of platform.}

\item{sample.wise}{Logical value, whether to extract sample-wise information. Default: FALSE.}
}
\value{
A dataframe.
}
\description{
Extract GEO Study Information.
}
\examples{
# pf.obj = GEOobj(acce = "GSE147507", platform = "GPL28369")
# pf.info = ExtractGEOInfo(pf.obj)
}
