% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{ExtractGEOMeta}
\alias{ExtractGEOMeta}
\title{Extract Sample Metadata.}
\usage{
ExtractGEOMeta(pf.obj)
}
\arguments{
\item{pf.obj}{GEO object of platform.}
}
\value{
A dataframe.
}
\description{
Extract Sample Metadata.
}
\examples{
# pf.obj = GEOobj(acce = "GSE147507", platform = "GPL28369")
# pf.info = ExtractGEOMeta(pf.obj)
}
