% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_ATAC.R
\name{FindMotif}
\alias{FindMotif}
\title{Find Motif of Peaks from RNA-seq and Peak-related Data Integrated Results.}
\usage{
FindMotif(
  inte.res,
  peak.motif.key,
  peak.mode = c("consensus", "diff"),
  homer.motif.path = NULL,
  genome = "mm10",
  out.folder = NULL,
  other.paras = NULL
)
}
\arguments{
\item{inte.res}{RNA and peak-related data integration results.}

\item{peak.motif.key}{The key type of integrated results ("Type" column of \code{inte.res}) to find motif.}

\item{homer.motif.path}{The path to Homer's \code{findMotifsGenome.pl}. Default: NULL (conduct automatic detection).}

\item{genome}{Parameter for \code{findMotifsGenome.pl}, can be genome FASTA files or pre-build genome info. Default: mm10.}

\item{out.folder}{The output folder. Default: NULL (current directory).}

\item{other.paras}{Parameter for \code{findMotifsGenome.pl}, can be '-len 8,10,12 -size -100,50 -S 25'. Default: NULL.}
}
\value{
List contains known motif results of UPbATAC and DOWNbATAC.
}
\description{
Find Motif of Peaks from RNA-seq and Peak-related Data Integrated Results.
}
