% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_visualization.R
\name{GenePlot}
\alias{GenePlot}
\title{Gene expresion plot.}
\usage{
GenePlot(
  deobj,
  deres,
  group.key = NULL,
  ref.group = NULL,
  base = 10,
  fill.color = c("blue", "red"),
  fill.alpha = 0.6,
  gene.df = NULL,
  label.key = NULL,
  gene.num = 2,
  plot.col = 2,
  scales = "free",
  signif = "padj",
  signif.threashold = 0.05,
  l2fc.threashold = 1,
  legend.pos = "top"
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{deres}{Data frame contains all genes.}

\item{group.key}{Sample group information. When set NULL, select first column of metadata. Default: NULL.}

\item{ref.group}{Reference group name. When set NULL, select first element of groups. Default: NULL.}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed. Default: 10.}

\item{fill.color}{Color for box,}

\item{fill.alpha}{Opacity of a geom. Default: 0.6.}

\item{gene.df}{Gene data frame, at least contains Gene column. Default: NULL. When set NULL, use \code{gene.num}.}

\item{label.key}{Column name in \code{gene.df} or \code{deres} to use as gene plot title. Default: NULL. When set NULL, use Gene column.}

\item{gene.num}{Gene number to plot, choose according to log2FoldChange. When \code{gene.df} is set NULL, use this to determine genes to plot. Default: NULL.}

\item{plot.col}{Column number of final plot. Default: 2.}

\item{scales}{Scales same as \code{\link{facet_wrap}}.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threashold}{Significance threashold to get differentially expressed genes. Default: 0.05.}

\item{l2fc.threashold}{Log2 fold change threashold to get differentially expressed genes. Default: 1.}

\item{legend.pos}{Legend position. Default: top.}
}
\value{
A ggplot2 object.
}
\description{
Gene expresion plot.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
dds$condition <- relevel(dds$condition, ref = "WT")
dds <- DESeq(dds)
dds.results <- results(dds, contrast = c("condition", "KO", "WT"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
GenePlot(deobj = dds, deres = dds.results.ordered, group.key = "condition", ref.group = "WT", fill.color = c("red", "blue"), fill.alpha = 0.8, gene.num = 2, signif = "pvalue", l2fc.threashold = 0.3)
}
