% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetGeneLength}
\alias{GetGeneLength}
\title{Get Gene Length from GTF.}
\usage{
GetGeneLength(gtf.file, save = FALSE, out.file = NULL)
}
\arguments{
\item{gtf.file}{GTF file.}

\item{save}{Logical value, whether to save the dataframe.Default: FALSE.}

\item{out.file}{File to save gene length, used when \code{save} is TRUE. Default: NULL.}
}
\value{
A dataframe or NULL.
}
\description{
Get Gene Length from GTF.
}
\examples{
# GetGeneLength(gtf.file, save = TRUE)
}
