% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IDConversion}
\alias{IDConversion}
\title{Gene ID Conversion.}
\usage{
IDConversion(
  deres,
  gene.type = c("ENSEMBL", "ENTREZID", "SYMBOL"),
  org.db = "org.Mm.eg.db",
  gene.map = NULL,
  sort.key = "log2FoldChange"
)
}
\arguments{
\item{deres}{Data frame contains all genes.}

\item{gene.type}{Gene name type. Chosen from ENSEMBL, ENTREZID,SYMBOL. Default: ENSEMBL.}

\item{org.db}{Organism database. Default: org.Mm.eg.db.}

\item{gene.map}{Use data frame instead of \code{org.db} to conduct gene id conversion, first column should be Gene. Default: NULL.}

\item{sort.key}{Column to sort the results. Default: log2FoldChange.}
}
\value{
Data frame contains gene id conversion results.
}
\description{
Gene ID Conversion.
}
\examples{
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
IDConversion(count.matrix, gene.type = "ENSEMBL", sort.key = NULL)
}
