% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferRegulator.R
\name{InferRegulator}
\alias{InferRegulator}
\title{Identify TFs from Gene Expression Reults.}
\usage{
InferRegulator(
  genes,
  control.genes,
  method = c("ChEA3", "BART2", "TFEA.ChIP"),
  species = c("Human", "Mouse"),
  projet.name = NULL,
  library = c("meanRank", "topRank", "GTEx", "ReMap", "Enrichr", "ENCODE", "ARCHS4",
    "Literature"),
  bart2.path = NULL,
  filter.threshold = NULL
)
}
\arguments{
\item{genes}{Input genes, can be a set of differentially expressed genes. When \code{method} is "ChEA3" or "BART2",
the genes should be SYMBOL of human or mouse. When \code{method} is "TFEA.ChIP", the genes can be
SYMBOL/ENSEMBL/ENTREZ of human.}

\item{control.genes}{Input genes of human, can be a set of non-differentially expressed genes.
Used when \code{method} is "TFEA.ChIP".}

\item{method}{The method used to identify TFs, choose from "ChEA3", "BART2", "TFEA.ChIP". Default: "ChEA3".}

\item{species}{The species of input gene set. Choose from "Human" and "Mouse" when \code{method} is "ChEA3" or "BART2",
"Human" when \code{method} is "TFEA.ChIP". Default: "Human".}

\item{projet.name}{The name of the project used as Query Name. Default: NULL (DEbPeak).}

\item{library}{The output library type, choose from "meanRank", "topRank", "GTEx", "ReMap",
"Enrichr", "ENCODE", "ARCHS4", "Literature". Used when \code{method} is "ChEA3". Default: "meanRank".}

\item{bart2.path}{BART2 path. Default: NULL (conduct automatic detection).}

\item{filter.threshold}{The threshold used to filter the results. When \code{method} is "ChEA3", filter on Score column.
When \code{method} is "BART2", filter on pvalue column. When \code{method} is "TFEA.ChIP", filter on pVal column.}
}
\value{
Dataframe contains identified TFs.
}
\description{
Identify TFs from Gene Expression Reults.
}
\examples{
# library(DEbPeak)
# InferRegulator(genes = c("SMAD9","FOXO1","MYC","STAT1",'STAT3',"SMAD3"), method = "ChEA3", species = "Human")
}
