% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_Peak.R
\name{InteDiffQuad}
\alias{InteDiffQuad}
\title{Create Quadrant Diagram for Two Differential Analysis Integration Results.}
\usage{
InteDiffQuad(
  inte.res,
  inte.type = c("DEbPeak", "PeakbPeak", "DEbDE"),
  point.alpha = 0.6,
  point.size.vec = c(2, 4),
  f1.l2fc.threshold = 0,
  f2.l2fc.threshold = 0,
  linetype = 2,
  point.color.vec = c("grey", "red", "blue", "grey"),
  legend.pos = "top",
  show.corr = TRUE,
  label.num = NULL,
  label.df = NULL,
  label.color = NULL
)
}
\arguments{
\item{inte.res}{Integration results, can be output of \code{DEbPeak}, \code{PeakbPeak}, \code{DEbDE}.}

\item{inte.type}{The integration type, choose from "DEbDE", "PeakbPeak", "DEbPeak". Default: "DEbPeak".}

\item{point.alpha}{Opacity of a geom. Default: 0.6.}

\item{point.size.vec}{Point size for regular and(or) labeled points (specified by \code{label.df} or \code{label.num}).
Default: 2 for regular and 4 for labeled points.}

\item{f1.l2fc.threshold}{Log2 fold change threshold for file1 to get differential analysis results. Default: 0.}

\item{f2.l2fc.threshold}{Log2 fold change threshold for file2 to get differential analysis results. Default: 0.}

\item{linetype}{Threshold linetype. Default: 2.}

\item{point.color.vec}{Point color for Down_Up, Up_Up, Down_Down, Up_Down.
Default: grey for Down_Up, red for Up_Up, blue for Down_Down and grey for Up_Down.}

\item{legend.pos}{Legend position. Default: top.}

\item{show.corr}{Logical value, whether to add pearson correlation coefficient and its significance. Default: TRUE.}

\item{label.num}{Gene number to label, according to RNA_log2FoldChange and Peak_log2FoldChange. When \code{label.df} is set NULL,
use this to determine genes to label. Default: NULL.}

\item{label.df}{Label data frame, at least contains Gene column. Default: NULL(use \code{label.num}).
When provided, the second column should not be in \code{de.peak}.}

\item{label.color}{Color for labels. Default: NULL (black).}
}
\value{
A ggplot2 object.
}
\description{
Create Quadrant Diagram for Two Differential Analysis Integration Results.
}
\examples{
# RNA-seq and RNA-seq
# de.de.label.df = data.frame(Gene = c("ENSDARG00000007396", "ENSDARG00000010729", "ENSDARG00000002194", "ENSDARG00000002587"))
# de.de.quad = InteDiffQuad(inte.res = de.de, inte.type = "DEbDE", f1.l2fc.threshold = 0.6, f2.l2fc.threshold = 0.6,
#                           show.corr = FALSE, label.df = de.de.label.df)
# Peak and Peak
# peak.peak.label.df = data.frame(Gene = c("Aak1", "Adam19", "Oaf", "Nsg2"))
# peak.peak.quad = InteDiffQuad(inte.res = atac.atac, inte.type = "PeakbPeak", f1.l2fc.threshold = 0,
#                               f2.l2fc.threshold = 0.5, show.corr = TRUE, label.df = peak.peak.label.df)
# RNA-seq and Peak
# de.peak.label.df = data.frame(Gene = c("Ccl3", "Ccl5", "Cd28", "Cx3cr1", "Prdm1", "Tcf7", "Slamf6", "Id3", "Cxcr5"))
# de.peak.quad = InteDiffQuad(inte.res = debatac.res, inte.type = "DEbPeak", f1.l2fc.threshold = 0,
#                             f2.l2fc.threshold = 0, show.corr = TRUE, label.df = de.peak.label.df)
}
