% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_Peak.R
\name{InteVenn}
\alias{InteVenn}
\title{Create Venn Diagram for Two Differential Analysis Integration Results.}
\usage{
InteVenn(
  inte.res,
  inte.type = c("DEbPeak", "PeakbPeak", "DEbDE"),
  peak.type = c("ChIP", "ATAC", "Peak"),
  peak.mode = c("consenus", "diff"),
  gene.col = c("geneId", "ENSEMBL", "SYMBOL"),
  ...
)
}
\arguments{
\item{inte.res}{Integration results, can be output of \code{DEbPeak}, \code{PeakbPeak}, \code{DEbDE}.}

\item{inte.type}{The integration type, choose from "DEbDE", "PeakbPeak", "DEbPeak". Default: "DEbPeak".}

\item{peak.type}{Used when \code{inte.type} is "DEbPeak". The source of peaks, chosen from ATAC, ChIP and Peak (ChIP and ATAC). Default: ChIP.}

\item{peak.mode}{Used when \code{inte.type} is "DEbPeak" or "PeakbPeak". The source of peak results, choose from consenus (peak annotation) and diff (differential expression analysis).
Default: consenus.}

\item{gene.col}{Used when \code{inte.type} is "DEbPeak" and \code{peak.type} is Peak or \code{peak.mode} is "consenus" in \code{peak.type} ATAC or ChIP.
Column of \code{inte.res} contains genes. Same as \code{merge.key} in \code{\link{DEbPeak}}.}

\item{...}{Parameters for \code{\link{ggvenn}}.}
}
\value{
A ggplot2 object.
}
\description{
Create Venn Diagram for Two Differential Analysis Integration Results.
}
\examples{
library(DEbPeak)
#### RNA-seq and RNA-seq
rna.diff.file <- system.file("extdata", "RA_RNA_diff.txt", package = "DEbPeak")
de1.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
de2.res <- read.table(file = rna.diff.file, header = TRUE, sep = "\t")
# use same file as example
de.de <- DEbDE(de1.res = de1.res, de2.res = de2.res, de1.l2fc.threshold = 0.5, de2.l2fc.threshold = 1)
de.de.venn <- InteVenn(inte.res = de.de, inte.type = "DEbDE", show_percentage = FALSE)
#### peak-related and peak-related
# ChIP-seq data
chip.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
chip.df <- GetConsensusPeak(peak.file = chip.file)
chip.anno <- AnnoPeak(
  peak.df = chip.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# ATAC-seq data
atac.file <- system.file("extdata", "debatac_peaks.bed", package = "DEbPeak")
atac.df <- GetConsensusPeak(peak.file = atac.file)
atac.anno <- AnnoPeak(
  peak.df = atac.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# integrate
chip.atac <- PeakbPeak(peak1.res = chip.anno$df, peak2.res = atac.anno$df, peak.mode = "consenus", peak.anno.key = "Promoter")
# functional enrichment
chip.atac.venn <- InteVenn(inte.res = chip.atac, inte.type = "PeakbPeak", peak.mode = "consenus", show_percentage = FALSE)
#### RNA-seq and peak-related
library(DESeq2)
# ChIP-Seq data
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# RNA-Seq data
count.file <- system.file("extdata", "debchip_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "debchip_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
# create DESeqDataSet object
dds <- DESeq2::DESeqDataSetFromMatrix(
  countData = count.matrix, colData = meta.info,
  design = ~condition
)
# set control level
dds$condition <- relevel(dds$condition, ref = "NF")
# conduct differential expressed genes analysis
dds <- DESeq(dds)
# extract results
dds.results <- results(dds, contrast = c("condition", "RX", "NF"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
# Integrated with RNA-Seq
debchip.res <- DEbPeak(
  de.res = dds.results.ordered, peak.res = peak.anno[["df"]],
  peak.anno.key = "Promoter", merge.key = "SYMBOL"
)
debatac.res.venn <- InteVenn(
  inte.res = debatac.res, inte.type = "DEbPeak", peak.mode = "diff",
  peak.type = "ATAC", show_percentage = FALSE
)
}
