% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{LoadingGO}
\alias{LoadingGO}
\title{GO enrichment on PC loading genes.}
\usage{
LoadingGO(
  pca,
  pc = 1,
  gene.num = 200,
  out.folder = NULL,
  gene.type = c("ENSEMBL", "ENTREZID", "SYMBOL"),
  go.type = c("ALL", "BP", "MF", "CC"),
  enrich.pvalue = 0.05,
  enrich.qvalue = 0.05,
  org.db = "org.Mm.eg.db",
  padj.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  show.term = 15,
  str.width = 30,
  plot.resolution = 300,
  plot.width = 7,
  plot.height = 9,
  save = TRUE
)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{pc}{Selected PC. Default: 1.}

\item{gene.num}{Gene number to export for every PC. Default: 200.}

\item{out.folder}{Folder to save enrichment results. Default: wording directory.}

\item{gene.type}{Gene name type. Chosen from ENSEMBL, ENTREZID,SYMBOL. Default: ENSEMBL.}

\item{go.type}{GO enrichment type, chosen from ALL, BP, MF, CC. Default: ALL.}

\item{enrich.pvalue}{Cutoff value of pvalue. Default: 0.05.}

\item{enrich.qvalue}{Cutoff value of qvalue. Default: 0.05.}

\item{org.db}{Organism database. Default: org.Mm.eg.db.}

\item{padj.method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: BH.}

\item{show.term}{Number of enrichment term to show. Default: 15.}

\item{str.width}{Length of enrichment term in plot. Default: 30.}

\item{plot.resolution}{Resolution of plot. Default: 300.}

\item{plot.width}{The width of plot. Default: 7.}

\item{plot.height}{The height of plot. Default: 9.}

\item{save}{Logical value, whether to save all results. Default: TRUE.}
}
\value{
If \code{save} is TRUE, return NULL (all results are in \code{out.folder}), else retutn list contains all results.
}
\description{
GO enrichment on PC loading genes.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
LoadingGO(pca_res, gene.type = "ENSEMBL", go.type = "BP", padj.method = "BH", save = TRUE)
}
