% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{LoadingPlot}
\alias{LoadingPlot}
\title{PCA loading plot.}
\usage{
LoadingPlot(
  pca,
  deobj = NULL,
  type = c("bar", "heat"),
  pc = 1:5,
  gene.num = 10,
  ncol = 2
)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{deobj}{Object created by DESeq2 or edgeR.}

\item{type}{loading plot type, chosen from bar, heat. Default: bar.}

\item{pc}{Specify PC to export genes. Default: 1:5.}

\item{gene.num}{Gene number to export for every PC. Default: 10.}

\item{ncol}{Column of final plots. Default: 2.}
}
\value{
Loading plot.
}
\description{
PCA loading plot.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
LoadingPlot(pca = pca_res, type = "bar")
LoadingPlot(pca = pca_res, deobj = dds, type = "heat")
}
