% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_enrichment.R
\name{MotifCompare}
\alias{MotifCompare}
\title{Map Motifs against a Motif Database.}
\usage{
MotifCompare(
  motif.folder,
  known.motif,
  tomtom.path = NULL,
  tomtom.paras = "-no-ssc -min-overlap 5 -dist pearson -evalue -thresh 10.0",
  out.folder = NULL
)
}
\arguments{
\item{motif.folder}{Folder contains \code{\link{MotifDiscovery}} results.}

\item{known.motif}{Motif database, should be in meme format.}

\item{tomtom.path}{The path to MEME's tomtom. Default: NULL (conduct automatic detection).}

\item{tomtom.paras}{Parameter for MEME's tomtom. Default: "-no-ssc -min-overlap 5 -dist pearson -evalue -thresh 10.0".}

\item{out.folder}{Output folder. Default: NULL (current working directory).}
}
\value{
Dataframe contains motif mapping results.
}
\description{
Map Motifs against a Motif Database.
}
\examples{
# library(DEbPeak)
# motif.cmp = MotifCompare(motif.folder = "/path/to/MotifDiscovery/out", known.motif = "/path/to/database.meme",
#                          tomtom.path = '/path/to/tomtom', out.folder = "/path/to/output")
}
