% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_enrichment.R
\name{MotifDiscovery}
\alias{MotifDiscovery}
\title{de novo Motif Discovery with STREME.}
\usage{
MotifDiscovery(
  peak.df = NULL,
  diff.peak = NULL,
  inte.res = NULL,
  peak.motif.key = NULL,
  genome,
  signif = "padj",
  signif.threshold = 0.05,
  l2fc.threshold = 0,
  peak.mode = c("consensus", "diff"),
  samtools.path = NULL,
  streme.path = NULL,
  streme.paras = "--nmotifs 50 --minw 8 --maxw 15 --thresh 0.05 --align center",
  out.folder = NULL
)
}
\arguments{
\item{peak.df}{Dataframe contains peaks, should contains chr, start, stop columns. Default: NULL.}

\item{diff.peak}{Dataframe contains differential peaks, the rownames should be peak annotations results. Default: NULL.}

\item{inte.res}{The Dataframe contains integrated results. \code{peak.df}, \code{diff.peak} and \code{inte.res}
cannot be empty at the same time. Dafault: NULL.}

\item{peak.motif.key}{The key type of integrated results ("Type" column of \code{inte.res}) or
regulation type of differential peaks (Up_regulated, Not_regulated, Down_regulated) to perfrom de novo motif discovery.
Used when \code{inte.res} is not NULL. Dafault: NULL.}

\item{genome}{The genome fasta path.}

\item{signif}{Significance criterion to get differential peaks. Used when \code{diff.peak} is not NULL.
For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threshold}{Significance threshold to get differential peaks. Used when \code{diff.peak} is not NULL. Default: 0.05.}

\item{l2fc.threshold}{Log2 fold change threshold to get differential peaks. Used when \code{diff.peak} is not NULL. Default: 0.}

\item{peak.mode}{The source of \code{inte.res} results, choose from consensus (consensus mode) and diff (differential analysis).
Default: consensus.}

\item{samtools.path}{The path to samtools (the version should >=1.9). Default: NULL (conduct automatic detection).}

\item{streme.path}{The path to MEME's streme. Default: NULL (conduct automatic detection).}

\item{streme.paras}{Parameter for MEME's streme.
Default: "--nmotifs 50 --minw 8 --maxw 15 --thresh 0.05 --align center".}

\item{out.folder}{Output folder. Default: NULL (current working directory).}
}
\value{
NULL.
}
\description{
de novo Motif Discovery with STREME.
}
\examples{
# library(DEbPeak)
# # de novo motif discovery with peak sets
# peak.file = system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
# peak.df = GetConsensusPeak(peak.file = peak.file)
# MotifDiscovery(peak.df = peak.df, genome = '/path/to/genome.fa',
#                streme.path = "/path/to/streme", samtools.path = "/path/to/samtools", out.folder = "/path/to/output")
# # de novo motif discovery with differential peaks
# MotifDiscovery(diff.peak = dds.peak.results.ordered, peak.motif.key = "Up_regulated",
#                genome = '/path/to/genome.fa', streme.path = "/path/to/streme",
#                samtools.path = "/path/to/samtools", out.folder = "/path/to/output")
# # de novo motif discovery with integrated results
# MotifDiscovery(inte.res = debatac.res, peak.motif.key = "Up_Up",
#                genome = '/path/to/genome.fa', peak.mode = "diff", streme.path = "/path/to/streme",
#                samtools.path = "/path/to/samtools", out.folder = "/path/to/output")
}
