% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motif_enrichment.R
\name{MotifEnrich}
\alias{MotifEnrich}
\title{Motif Enrichment for Differentially Accessible/Binding Peaks.}
\usage{
MotifEnrich(
  de.res,
  reg.type,
  signif = "padj",
  signif.threshold = 0.05,
  l2fc.threshold = 0,
  homer.motif.path = NULL,
  genome = "mm10",
  out.folder = NULL,
  other.paras = NULL
)
}
\arguments{
\item{de.res}{Differential analysis results of peak-related data.}

\item{reg.type}{The regulation type, choose from Up_regulated, Not_regulated, Down_regulated.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threshold}{Significance threshold to get differentially expressed results. Default: 0.05.}

\item{l2fc.threshold}{Log2 fold change threshold to get differentially expressed results. Default: 0.}

\item{homer.motif.path}{The path to Homer's \code{findMotifsGenome.pl}. Default: NULL.}

\item{genome}{Parameter for \code{findMotifsGenome.pl}, can be genome FASTA files or pre-build genome info. Default: mm10.}

\item{out.folder}{The output folder. Default: NULL (current directory).}

\item{other.paras}{Parameter for \code{findMotifsGenome.pl}, can be '-len 8,10,12 -size -100,50 -S 25'. Default: NULL.}
}
\value{

}
\description{
Motif Enrichment for Differentially Accessible/Binding Peaks.
}
\examples{
# dds.peak.results.ordered is differential analysis of peak-related data.
# peak.motif = MotifEnrich(de.res = dds.peak.results.ordered, reg.type = "Up_regulated",
#                          homer.motif.path = '~/anaconda3/bin/findMotifsGenome.pl',
#                          out.folder = "/path/to/out/folder", other.paras = "-len 8,10,12 -size -100,50 -S 25")
}
