% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_Peak.R
\name{NetViz}
\alias{NetViz}
\title{Visualization for Enhancer-Gene Network.}
\usage{
NetViz(
  inte.res,
  type,
  whole = TRUE,
  gene = NULL,
  peak = NULL,
  labels = NULL,
  show.all.labels = FALSE,
  seed = 1000,
  edge.width = 0.6,
  edge.color = "black",
  edge.alpha = 0.8,
  node.size = 3,
  node.color = c("red", "blue"),
  node.alpha = 0.5,
  node.label.size = 2.5,
  node.label.color = "navy",
  node.label.len = 0.5
)
}
\arguments{
\item{inte.res}{Dataframe contains integrated results.}

\item{type}{The type of integrated results ("Type" column of \code{inte.res}) to perform visualization.}

\item{whole}{Logical value, whether to visualize all interactions. Default: TRUE.}

\item{gene}{Genes used to extract subset of the whole network. Default: NULL.}

\item{peak}{Peaks used to extract subset of the whole network. Default: NULL.}

\item{labels}{Genes/Peaks used to label on the plot. Default: NULL.}

\item{show.all.labels}{Logical value, whether to show all labels, useful when extracting subset of the whole network. Default: FALSE.}

\item{seed}{Seed.}

\item{edge.width}{The width of the edge. Default: 0.6.}

\item{edge.color}{The color of the edge. Default: "black".}

\item{edge.alpha}{The alpha of the edge. Default: 0.8.}

\item{node.size}{The size of the node. Default: 3.}

\item{node.color}{The color vector for enhancer and gene. Default: c("red", "blue").}

\item{node.alpha}{The alpha of the node color. Default: 0.6.}

\item{node.label.size}{The size of the node label. Default: 2.5.}

\item{node.label.color}{The color of the node label. Default: "navy".}

\item{node.label.len}{The line length of the node label. Default: 0.5.}
}
\value{
A ggplot2 object.
}
\description{
Visualization for Enhancer-Gene Network.
}
\examples{
# # whole network
# NetViz(inte.res = test, whole = TRUE, type = "Up_Up",
#        labels = c("Ripor2", "C4b", "Ccl12", "Fcgr2b", "Ly86", "Ccl5", "Ccr5"), seed = 1001)
# # subset with gene and peak
# NetViz(inte.res = test, whole = FALSE, type = "Up_Up",
#        gene = c("Ripor2", "C4b", "Ccl12", "Fcgr2b", "Ly86", "Ccl5", "Ccr5"),
#        peak = c("15:74931074-74931694|Ly6e|P", "9:124012191-124012495|Ccr3|P", "7:141122254-141122774|Ptdss2|P", "1:170985190-170986301|Fcgr2b|P"),
#        show.all.labels = TRUE, seed = 1001)
# # single gene
# NetViz(inte.res = test, whole = FALSE, type = "Up_Up",
#        gene = c("Ripor2", "C4b", "Ccl12", "Fcgr2b", "Ly86", "Ccl5", "Ccr5"),
#        show.all.labels = TRUE, seed = 1001)
# # single peak
# NetViz(inte.res = test, whole = FALSE, type = "Up_Up",
#        peak = c("15:74931074-74931694|Ly6e|P", "9:124012191-124012495|Ccr3|P", "7:141122254-141122774|Ptdss2|P", "1:170985190-170986301|Fcgr2b|P"),
#        show.all.labels = TRUE, seed = 1001)
}
