% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{OutlierDetection}
\alias{OutlierDetection}
\title{Detect outlier with robust PCA}
\usage{
OutlierDetection(
  deobj,
  var.genes = NULL,
  remove.sample = NULL,
  transform.method = c("rlog", "vst", "ntd"),
  rpca.method = c("PcaGrid", "PcaHubert"),
  k = 2,
  ...
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{var.genes}{Select genes with larger variance for PCA analysis. Default: all genes.}

\item{remove.sample}{Sample(s) to remove. Default: NULL.}

\item{transform.method}{Data transformation methods, chosen from rlog, vst and ntd. Default: rlog.}

\item{rpca.method}{Robust PCA method, chosen from \code{\link{PcaGrid}}, \code{\link{PcaHubert}}. Default: PcaGrid.}

\item{k}{Number of principal components to compute, for \code{\link{PcaGrid}}, \code{\link{PcaHubert}}. Default: 2.}

\item{...}{Parameter for \code{\link{PcaGrid}}, \code{\link{PcaHubert}}.}
}
\value{
A ggplot2 object
}
\description{
Detect outlier with robust PCA
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
outlier.res <- OutlierDetection(deobj = dds, var.genes = NULL, transform.method = "rlog")
outlier.res$outlier
outlier.res$plot
}
