% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseGEO.R
\name{ParseGEO}
\alias{ParseGEO}
\title{Parse GEO Data.}
\usage{
ParseGEO(acce, platform, supp.idx = 1, down.supp = FALSE, ...)
}
\arguments{
\item{acce}{GEO accession number.}

\item{platform}{Platform information/field.}

\item{supp.idx}{The index of supplementary files to download. This should be consistent with \code{platform}. Default: 1.}

\item{down.supp}{Logical value, whether to download supplementary files to create count matrix. If TRUE, always
download supplementary files. If FALSE, use \code{ExpressionSet} (If contains non-integer or emoty,
download supplementary files automatically). Default: FALSE.}

\item{...}{Parameters for \code{\link{getGEO}}.}
}
\value{
List contains GEO object of platform, study information, raw count matrix and metadata.
}
\description{
Parse GEO Data.
}
\examples{
# GSE149838.list = ParseGEO(acce = "GSE149838", platform = "GPL21626")
# GPL28369 is consistent with first supplementary file
# GSE147507.list = ParseGEO(acce = "GSE147507", platform = "GPL28369", supp.idx = 1)
# GSE122774.list = ParseGEO(acce = "GSE122774", platform = "GPL17021")
}
