% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{PCA}
\alias{PCA}
\title{Calculate PCA.}
\usage{
PCA(
  deobj,
  var.genes = NULL,
  remove.sample = NULL,
  transform.method = c("rlog", "vst", "ntd")
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{var.genes}{Select genes with larger variance for PCA analysis. Default value is NULL, using all genes.}

\item{remove.sample}{Sample(s) to remove. Default: NULL.}

\item{transform.method}{Data transformation methods, chosen from rlog, vst and ntd. Default: rlog.}
}
\value{
List suitable for PCAtools.
}
\description{
Calculate PCA.
}
\examples{
library(DESeq2)
library(DEbPeak)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
}
