% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{PCA3D}
\alias{PCA3D}
\title{3D PCA plot.}
\usage{
PCA3D(
  pca,
  x = "PC1",
  y = "PC2",
  z = "PC3",
  color.key = NULL,
  color = NULL,
  cex = 1.5,
  col = "black",
  ticktype = "detailed",
  bty = "g",
  box = TRUE,
  theta = 140,
  phi = 20,
  d = 3,
  colkey = FALSE,
  label.size = 0.5,
  legend.pos = "bottom",
  legend.horiz = TRUE,
  legend.inset = -0.1,
  legend.bg = "white",
  legend.bty = "n",
  legend.ncol = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{x}{The principal component to display on the x axis. Default: PC1.}

\item{y}{The principal component to display on the y axis. Default: PC2.}

\item{z}{The principal component to display on the z axis. Default: PC3.}

\item{color.key}{Group to color the points. When set NULL, select first column of metadata. Default: NULL.}

\item{color}{Colors used. When set NULL, select color from Set1 of \code{RColorBrewer} package. Default: NULL.}

\item{cex}{Point size. Default: 1.5.}

\item{col}{Point edge color. Default: black.}

\item{ticktype}{Axis ticktype, chosen from: detailed and simple. Default: detailed.}

\item{bty}{The type of the box, the default draws only the back panels,
chosen from: f, b, b2, g, bl, bl2, u, n. Default: g.}

\item{box}{Should the bounding box for the surface be displayed. Default: TRUE.}

\item{theta}{Angles defining the viewing direction. \code{theta} gives the azimuthal direction. Default: 140.}

\item{phi}{Angles defining the viewing direction. \code{phi} gives the colatitude direction. Default: 20.}

\item{d}{a value which can be used to vary the strength of the perspective transformation.
Values of d greater than 1 will lessen the perspective effect and
values less and 1 will exaggerate it. Default: 3.}

\item{colkey}{\code{\link{colkey}}. Default: FALSE.}

\item{label.size}{Label font size. Default: 0.5.}

\item{legend.pos}{Legend position. Default: bottom.}

\item{legend.horiz}{Logical; if TRUE, set the legend horizontally rather than
vertically (specifying horiz overrides the ncol specification). Default: TRUE.}

\item{legend.inset}{Inset distance(s) from the margins as a fraction of the plot region when
legend is placed by keyword. Default: -0.1.}

\item{legend.bg}{The background color for the legend box.
Note that this is only used if bty != "n". Default: white.}

\item{legend.bty}{The type of box to be drawn around the legend,
chosen from 0 and n. Default: n.}

\item{legend.ncol}{The number of columns in which to set the legend items.
Default: NULL (set the legend horizontally).}

\item{main}{Title of the plot. Default: NULL.}

\item{...}{Parameters for \code{\link{scatter3D}}.}
}
\value{
3D PCA plot.
}
\description{
3D PCA plot.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
PCA3D(pca = pca_res, color.key = "condition", main = "3D PCA")
}
