% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{PCABasic}
\alias{PCABasic}
\title{Generated PCA baisc plots, including screen plot, biplot and pairs plot.}
\usage{
PCABasic(
  pca,
  x = "PC1",
  y = "PC2",
  explain.threashold = 90,
  loading.num = 5,
  loading.label.size = 3,
  loading.label.color = "red",
  colby = NULL,
  pair.pc = 5,
  pair.label.size = 14,
  legend.pos = c("right", "bottom", "top", "left", "none")
)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{x}{The principal component to display on the x axis. Default: PC1.}

\item{y}{The principal component to display on the y axis. Default: PC2.}

\item{explain.threashold}{The threashold of explained variance. Default: 90.}

\item{loading.num}{Select loading gene number based on absolute ordered variable loading for each PC in the biplot. Default: 5.}

\item{loading.label.size}{Size of loading label. Default: 3.}

\item{loading.label.color}{Color of loading label. Default: red.}

\item{colby}{Group information to color samples. Default: NULL.}

\item{pair.pc}{The principal components to be included in the plot. Default: 5.}

\item{pair.label.size}{Size of p rincipal component label. Default: 14.}

\item{legend.pos}{Position of legend ('top', 'bottom', 'left', 'right', 'none'). Default: right.}
}
\value{
A ggplot2 object.
}
\description{
Generated PCA baisc plots, including screen plot, biplot and pairs plot.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
PCABasic(pca_res, colby = "condition")
}
