% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_annotation.R
\name{PeakAnnoPie}
\alias{PeakAnnoPie}
\title{Create Peak Annotation Pie Plot with ggplot2.}
\usage{
PeakAnnoPie(peak.obj, ...)
}
\arguments{
\item{peak.obj}{Peak annotation object, returned from \code{\link{annotatePeak}}.}

\item{...}{Parameters for \code{\link{ggpie}}.}
}
\value{
A ggplot2 object.
}
\description{
Create Peak Annotation Pie Plot with ggplot2.
}
\examples{
library(DEbPeak)
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
PeakAnnoPie(peak.anno$anno.obj)
}
