% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_bind_peak.R
\name{PeakbPeak}
\alias{PeakbPeak}
\title{Integrate Two Peak Annotation/Differential Analysis Results.}
\usage{
PeakbPeak(
  peak1.res,
  peak2.res,
  peak.mode = c("consenus", "diff"),
  peak.anno.key = c("Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream",
    "Distal Intergenic", "All"),
  peak1.signif = "padj",
  peak1.signif.threshold = 0.05,
  peak1.l2fc.threshold = 1,
  peak2.signif = "padj",
  peak2.signif.threshold = 0.05,
  peak2.l2fc.threshold = 1
)
}
\arguments{
\item{peak1.res}{Peak1 dataframe contains all peak annotation (\code{peak.mode} is consenus) or
differential analysis results of peak-related data (\code{peak.mode} is diff).}

\item{peak2.res}{Peak2 dataframe contains all peak annotation (\code{peak.mode} is consenus) or
differential analysis results of peak-related data (\code{peak.mode} is diff).}

\item{peak.mode}{The source of peak results, choose from consenus (peak annotation) and diff (differential analysis).
Default: consenus.}

\item{peak.anno.key}{Peak location, chosen from "Promoter", "5' UTR", "3' UTR", "Exon", "Intron", "Downstream", "Distal Intergenic","All".
Used when \code{peak.mode} is consenus. Default: "Promoter".}

\item{peak1.signif}{Used when \code{peak.mode} is diff. Significance criterion for peak-associated results for Peak1. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{peak1.signif.threshold}{Used when \code{peak.mode} is diff. Significance threshold for peak-associated results to get differentially accessible/binding peaks for Peak1. Default: 0.05.}

\item{peak1.l2fc.threshold}{Used when \code{peak.mode} is diff. Log2 fold change threshold for peak-associated results to get differentially accessible/binding peaks for Peak1. Default: 1.}

\item{peak2.signif}{Used when \code{peak.mode} is diff. Significance criterion for peak-associated results for Peak2. For DESeq2 results, can be chosen from padj, pvalue.
For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{peak2.signif.threshold}{Used when \code{peak.mode} is diff. Significance threshold for peak-associated results to get differentially accessible/binding peaks for Peak2. Default: 0.05.}

\item{peak2.l2fc.threshold}{Used when \code{peak.mode} is diff. Log2 fold change threshold for peak-associated results to get differentially accessible/binding peaks for Peak2. Default: 1.}
}
\value{
Dataframe contains integration results. When \code{peak.mode} is 'diff', the 'Type' column contains "Down_Up", "Up_Up", "Down_Down", "Up_Down", "Peak1_Up", "Peak1_Down", "Peak2_Up", "Peak2_Down".
When \code{peak.mode} is 'consenus', the 'Type' column contains "Common", "Peak1", "Peak2".
}
\description{
Integrate Two Peak Annotation/Differential Analysis Results.
}
\examples{
library(DEbPeak)
# ChIP-seq data
chip.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
chip.df <- GetConsensusPeak(peak.file = chip.file)
chip.anno <- AnnoPeak(
  peak.df = chip.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# ATAC-seq data
atac.file <- system.file("extdata", "debatac_peaks.bed", package = "DEbPeak")
atac.df <- GetConsensusPeak(peak.file = atac.file)
atac.anno <- AnnoPeak(
  peak.df = atac.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# integrate
chip.atac <- PeakbPeak(peak1.res = chip.anno$df, peak2.res = atac.anno$df, peak.mode = "consenus", peak.anno.key = "Promoter")
}
