% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_bind_peak.R
\name{PeakbPeakFE}
\alias{PeakbPeakFE}
\title{GO Enrichment on Two Peak Annotation/Differential Analysis Integration Results.}
\usage{
PeakbPeakFE(
  peak.peak,
  peak.fe.key,
  out.folder = NULL,
  gene.type = c("ENSEMBL", "ENTREZID", "SYMBOL"),
  go.type = c("ALL", "BP", "MF", "CC"),
  enrich.pvalue = 0.05,
  enrich.qvalue = 0.05,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  padj.method = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  show.term = 15,
  str.width = 30,
  plot.resolution = 300,
  plot.width = 7,
  plot.height = 9,
  save = TRUE
)
}
\arguments{
\item{peak.peak}{Dataframe contains integrated results of two peak annotation/differential analysis.}

\item{peak.fe.key}{The key type of integrated results ("Type" column of \code{peak.peak}) to perform functional enrichment.}

\item{out.folder}{Folder to save enrichment results. Default: wording directory.}

\item{gene.type}{Gene name type. Chosen from ENSEMBL, ENTREZID,SYMBOL. Default: ENSEMBL.}

\item{go.type}{GO enrichment type, chosen from ALL, BP, MF, CC. Default: ALL.}

\item{enrich.pvalue}{Cutoff value of pvalue. Default: 0.05.}

\item{enrich.qvalue}{Cutoff value of qvalue. Default: 0.05.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{padj.method}{One of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none". Default: BH.}

\item{show.term}{Number of enrichment term to show. Default: 15.}

\item{str.width}{Length of enrichment term in plot. Default: 30.}

\item{plot.resolution}{Resolution of plot. Default: 300.}

\item{plot.width}{The width of plot. Default: 7.}

\item{plot.height}{The height of plot. Default: 9.}

\item{save}{Logical value, whether to save all results. Default: TRUE.}
}
\value{
If \code{save} is TRUE, return NULL (all results are in \code{out.folder}), else retutn result dataframe.
}
\description{
GO Enrichment on Two Peak Annotation/Differential Analysis Integration Results.
}
\examples{
library(DEbPeak)
# ChIP-seq data
chip.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
chip.df <- GetConsensusPeak(peak.file = chip.file)
chip.anno <- AnnoPeak(
  peak.df = chip.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# ATAC-seq data
atac.file <- system.file("extdata", "debatac_peaks.bed", package = "DEbPeak")
atac.df <- GetConsensusPeak(peak.file = atac.file)
atac.anno <- AnnoPeak(
  peak.df = atac.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)
# integrate
chip.atac <- PeakbPeak(peak1.res = chip.anno$df, peak2.res = atac.anno$df, peak.mode = "consenus", peak.anno.key = "Promoter")
# functional enrichment
chip.atac.fe <- PeakbPeakFE(
  peak.peak = chip.atac, peak.fe.key = "Common", gene.type = "SYMBOL",
  go.type = "BP", species = "Mouse", save = FALSE
)
}
