% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIP_peak_annotation.R
\name{PeakProfile}
\alias{PeakProfile}
\title{Create ChIP Peak Binding Profile.}
\usage{
PeakProfile(
  peak.df,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  gtf.file = NULL,
  weight.col = NULL,
  up.dist = 3000,
  down.dist = 3000,
  color = "red",
  conf = 0.95,
  up.rel = 0.2,
  down.rel = 0.2,
  by = c("gene", "transcript", "exon", "intron", "3UTR", "5UTR"),
  region.type = c("start_site", "end_site", "body"),
  nbin = NULL
)
}
\arguments{
\item{peak.df}{Dataframe contains all consensus peaks.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis",
"Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{gtf.file}{GTF file used to create TxDb object. Useful when specie you used is not available in \code{species}. Default: NULL.}

\item{weight.col}{The column name of weight. Parameter of \code{\link{peakHeatmap}},
\code{\link{plotAvgProf2}}, \code{\link{plotPeakProf2}}. Default: NULL.}

\item{up.dist}{The upstream distance from the TSS. Default: 3000bp.}

\item{down.dist}{The downstream distance from the TSS. Default: 3000bp.}

\item{color}{The color used for heatmap. Parameter of \code{\link{peakHeatmap}}. Default: red.}

\item{conf}{The confidence interval. Parameter of \code{\link{plotAvgProf2}}, \code{\link{plotPeakProf2}}. Default: 0.95.}

\item{up.rel}{The percentage of total length of body regions used as upstream distance from the TSS.
Parameter of \code{\link{plotPeakProf2}}. When set NULL, use \code{up.dist}. Default: 0.2.}

\item{down.rel}{The percentage of total length of body regions used as downstream distance from the TSS.
Parameter of \code{\link{plotPeakProf2}}. When set NULL, use \code{down.dist}. Default: 0.2.}

\item{by}{One of 'gene', 'transcript', 'exon', 'intron' , '3UTR' , '5UTR'. Parameter of \code{\link{plotPeakProf2}}. Default: 'gene'.}

\item{region.type}{One of "start_site", "end_site", "body". Parameter of \code{\link{plotPeakProf2}}. Default: 'start_site'.}

\item{nbin}{The amount of nbines. Parameter of \code{\link{plotPeakProf2}}. Default: NULL.}
}
\value{
List contains all profile plots.
}
\description{
Create ChIP Peak Binding Profile.
}
\examples{
library(DEbChIP)
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbChIP")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)
}
