% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{QCPCA}
\alias{QCPCA}
\title{PCA related functions used in quality control.}
\usage{
QCPCA(
  deobj,
  var.genes = NULL,
  remove.sample = NULL,
  transform.method = c("rlog", "vst", "ntd"),
  batch = NULL,
  raw.deobj = NULL,
  min.count = 10,
  colby = NULL,
  legend.pos = c("bottom", "top", "left", "right", "none"),
  outlier.detection = TRUE,
  rpca.method = c("PcaGrid", "PcaHubert"),
  k = 2,
  ...
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{var.genes}{Select genes with larger variance for PCA analysis. Default: all genes.}

\item{remove.sample}{Sample(s) to remove. Default: NULL.}

\item{transform.method}{Data transformation methods, chosen from rlog, vst and ntd. Default: rlog.}

\item{batch}{Batch column to conduct batch correction. Default value is NULL, do not conduct batch correction.}

\item{raw.deobj}{Object created by DESeq2 or edgeR before filtering with counts. Default: NULL.}

\item{min.count}{A feature is considered to be detected if the corresponding number of read counts is > \code{min.count}. By default, \code{min.count} = 10.}

\item{colby}{Group information to color samples. Default: NULL.}

\item{legend.pos}{Position of legend ('top', 'bottom', 'left', 'right', 'none'). Default: bottom.}

\item{outlier.detection}{Logical value. If TRUE, conduct outlier detection with robust PCA. Default: TRUE.}

\item{rpca.method}{Robust PCA method, chosen from \code{\link{PcaGrid}}, \code{\link{PcaHubert}}. Default: PcaGrid.}

\item{k}{Number of principal components to compute, for \code{\link{PcaGrid}}, \code{\link{PcaHubert}}. Default: 2.}

\item{...}{Parameter for \code{\link{PcaGrid}}, \code{\link{PcaHubert}}.}
}
\value{
List contains all step plots, final PCA results, final deobj.
}
\description{
PCA related functions used in quality control.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pcs_res <- QCPCA(
  deobj = dds, var.genes = NULL, remove.sample = NULL, transform.method = "rlog",
  outlier.detection = TRUE, rpca.method = "PcaGrid"
)
}
