% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_annotation.R
\name{ReadProfile}
\alias{ReadProfile}
\title{Generate Reads Profile with deepTools.}
\usage{
ReadProfile(
  bw.folder,
  up.dist = 3000,
  down.dist = 3000,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  gtf.file = NULL,
  compute.matrix.paras = "-p 2",
  plot.heatmap.paras = "--whatToShow 'heatmap and colorbar'",
  deeptools.path = NULL,
  out.folder = NULL,
  out.format = c("pdf", "png", "eps", "svg")
)
}
\arguments{
\item{bw.folder}{Folder contains bigWig files.}

\item{up.dist}{The upstream distance from the TSS. Default: 3000bp.}

\item{down.dist}{The downstream distance from the TSS. Default: 3000bp.}

\item{species}{Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus",
"Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".}

\item{gtf.file}{GTF file used to create TxDb object. Useful when specie you used is not available in \code{species}. Default: NULL.}

\item{compute.matrix.paras}{Parameters for deepTools \code{computeMatrix}. Default: "-p 2".}

\item{plot.heatmap.paras}{Parameters for deepTools \code{plotHeatmap}. Default: "--whatToShow 'heatmap and colorbar'".}

\item{deeptools.path}{Path contains \code{computeMatrix} and \code{plotHeatmap}. Default: NULL (conduct automatic detection).}

\item{out.folder}{Output folder. Default: NULL (current working directory).}

\item{out.format}{Output heatmap format. Choose from pdf, png, eps and svg. Default: pdf.}
}
\value{
NULL.
}
\description{
Generate Reads Profile with deepTools.
}
\examples{
# library(DEbPeak)
# ReadProfile(bw.folder = path/to/bigWigs, species = "Mouse", deeptools.path = "~/anaconda3/bin")
}
