% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferRegulator.R
\name{RunBART2}
\alias{RunBART2}
\title{Identify TFs from Gene Expression Reults with BART2.}
\usage{
RunBART2(
  genes,
  species = c("Human", "Mouse"),
  bart2.path = NULL,
  projet.name = NULL
)
}
\arguments{
\item{genes}{Input gene symbol of human or mouse, can be a set of differentially expressed genes.}

\item{species}{The species of input gene set. Choose from "Human" and "Mouse". Default: "Human".}

\item{bart2.path}{BART2 path. Default: NULL (conduct automatic detection).}

\item{projet.name}{The name of the project used as Query Name. Default: NULL (DEbPeak).}
}
\value{
Dataframe contains identified TFs.
}
\description{
Identify TFs from Gene Expression Reults with BART2.
}
\examples{
# library(DEbPeak)
# RunBART2(genes = c("SMAD9","FOXO1","MYC","STAT1",'STAT3',"SMAD3"), species = "Human",
#          bart2.path = "~/anaconda3/envs/bart2/bin/bart2")
}
