% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferRegulator.R
\name{RunChEA3}
\alias{RunChEA3}
\title{Identify TFs from Gene Expression Reults with ChEA3.}
\usage{
RunChEA3(
  genes,
  species = c("Human", "Mouse"),
  projet.name = NULL,
  library = c("meanRank", "topRank", "GTEx", "ReMap", "Enrichr", "ENCODE", "ARCHS4",
    "Literature")
)
}
\arguments{
\item{genes}{Input gene symbol of human or mouse, can be a set of differentially expressed genes.}

\item{species}{The species of input gene set. Choose from "Human" and "Mouse". Default: "Human".}

\item{projet.name}{The name of the project used as Query Name. Default: NULL (DEbPeak).}

\item{library}{The output library type, choose from "meanRank", "topRank", "GTEx", "ReMap",
"Enrichr", "ENCODE", "ARCHS4", "Literature". Default: "meanRank".}
}
\value{
Dataframe contains identified TFs.
}
\description{
Identify TFs from Gene Expression Reults with ChEA3.
}
\examples{
# library(DEbPeak)
# RunChEA3(genes = c("SMAD9","FOXO1","MYC","STAT1",'STAT3',"SMAD3"), species = "Human")
}
