% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferRegulator.R
\name{RunTFEA}
\alias{RunTFEA}
\title{Identify TFs from Gene Expression Reults with TFEA.ChIP.}
\usage{
RunTFEA(genes, control.genes)
}
\arguments{
\item{genes}{Input genes of human, can be a set of differentially expressed genes.}

\item{control.genes}{Input genes of human, can be a set of non-differentially expressed genes.}
}
\value{
Dataframe contains identified TFs with GSEA enrichment score.
}
\description{
Identify TFs from Gene Expression Reults with TFEA.ChIP.
}
\examples{
# genes = c("EGLN3","NFYA","ALS2","MYC","ARNT" )
# control.genes = c("MTX1", "OSTF1", "HDAC3", "MTG2", "OPA3")
# RunTFEA(genes = genes, control.genes = control.genes)
}
