% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_relation.R
\name{SampleRelation}
\alias{SampleRelation}
\title{Sample to sample relationship.}
\usage{
SampleRelation(
  deobj,
  gene = NULL,
  remove.sample = NULL,
  transform.method = c("rlog", "vst", "ntd"),
  anno.key = NULL,
  ...
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{gene}{Genes used to create sample relationship. Default: NULL.}

\item{remove.sample}{Sample(s) to remove. Default: NULL.}

\item{transform.method}{Data transformation methods, chosen from rlog, vst and ntd. Default: rlog.}

\item{anno.key}{Group to annotate samples. When set NULL, select first column of metadata. Default: NULL.}

\item{...}{Parameters for \code{\link{pheatmap}}.}
}
\value{
Plot of sample distances and sample pearson correlation.
}
\description{
Sample to sample relationship.
}
\examples{
library(DESeq2)
library(DEbPeak)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbPeak")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbPeak")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
SampleRelation(deobj = dds, transform.method = "rlog", anno.key = "condition")
}
