% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_visualization.R
\name{ScatterPlot}
\alias{ScatterPlot}
\title{ScatterPlot for Differentially Expressed Genes.}
\usage{
ScatterPlot(
  deobj,
  deres,
  group.key = NULL,
  ref.group = NULL,
  base = 10,
  signif = "padj",
  signif.threashold = 0.05,
  l2fc.threashold = 1,
  point.alpha = 0.6,
  point.size.vec = c(2, 4),
  linetype = 2,
  point.color.vec = c("blue", "grey", "red"),
  legend.pos = "top",
  label.num = NULL,
  label.df = NULL,
  label.key = NULL,
  label.color = NULL
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{deres}{Data frame contains all genes.}

\item{group.key}{Sample group information. When set NULL, select first column of metadata. Default: NULL.}

\item{ref.group}{Reference group name. When set NULL, select first element of groups. Default: NULL.}

\item{base}{A positive or complex number: the base with respect to which logarithms are computed. Default: 10.}

\item{signif}{Significance criterion. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.}

\item{signif.threashold}{Significance threashold to get differentially expressed genes. Default: 0.05.}

\item{l2fc.threashold}{Log2 fold change threashold to get differentially expressed genes. Default: 1.}

\item{point.alpha}{Opacity of a geom. Default: 0.6.}

\item{point.size.vec}{Point size for regular(DEG or non-DEG) and(or) labeled points. Default: 2 for regular and 4 for labeled points.}

\item{linetype}{Diagonal linetype. Default: 2.}

\item{point.color.vec}{Point color for Down, Not, Up regulated genes. Default: red for Up, grey for Not and blue for Down.}

\item{legend.pos}{Legend position. Default: top.}

\item{label.num}{Gene number to label, choose according to log2FoldChange. When \code{label.df} is set NULL, use this to determine genes to label. Default: NULL.}

\item{label.df}{Label data frame, at least contains Gene column. Default: NULL(use \code{label.num}). When provided,
the second column should not be in \code{deres}.}

\item{label.key}{Which column to use as label. Default: NULL (use rownames of \code{deres} or Gene column of \code{gene.df}).}

\item{label.color}{Color vector for labels. Default: NULL.}
}
\value{
A ggplot2 object.
}
\description{
ScatterPlot for Differentially Expressed Genes.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
dds$condition <- relevel(dds$condition, ref = "WT")
dds <- DESeq(dds)
dds.results <- results(dds, contrast = c("condition", "KO", "WT"))
dds.results.ordered <- dds.results[order(dds.results$log2FoldChange, decreasing = TRUE), ]
ScatterPlot(deobj = dds, deres = dds.results.ordered, group.key = "condition", ref.group = "WT", signif = "pvalue", l2fc.threashold = 0.3, label.num = 2, point.alpha = 0.8, label.color = c("purple", "green"))
}
