% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_enrichment_analysis.R
\name{VisGSEA}
\alias{VisGSEA}
\title{Visualize GSEA results.}
\usage{
VisGSEA(gsea.res, gs.id = NULL, ncol = 1, ...)
}
\arguments{
\item{gsea.res}{GSEA results, should be gseaResult object.}

\item{gs.id}{Gene set ID name used to create plot, can be single or multiple. Default: NULL.}

\item{ncol}{The plot column when \code{gs.id} is larger than 1. Default: 1.}

\item{...}{Parameters for \code{\link{gseaplot2}}.}
}
\value{
A ggplot2 or patchwork object.
}
\description{
Visualize GSEA results.
}
\examples{
# VisGSEA(gsea.res = gsea.results$gsea, gs.id = c("GOCC_CONTRACTILE_FIBER", "GOMF_SIGNALING_RECEPTOR_REGULATOR_ACTIVITY"))
}
