% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferRegulator.R
\name{VizRegulator}
\alias{VizRegulator}
\title{Visualize the Identified TFs.}
\usage{
VizRegulator(
  infer.res,
  method = c("ChEA3", "BART2", "TFEA.ChIP"),
  label.gene = NULL,
  point.alpha = 0.5,
  label.color = "red"
)
}
\arguments{
\item{infer.res}{The identidied result dataframe, can be the output of \code{\link{InferRegulator}}.}

\item{method}{The method used to identify TFs, choose from "ChEA3", "BART2", "TFEA.ChIP". Default: "ChEA3".}

\item{label.gene}{The highlight genes. Default: NULL.}

\item{point.alpha}{The point alpha of identify TFs. Default: 0.5.}

\item{label.color}{The text/point color for the highlight genes. Default: "red".}
}
\value{
A ggplot2 object.
}
\description{
Visualize the Identified TFs.
}
\examples{
# library(DEbPeak)
# infer.res = InferRegulator(genes = c("SMAD9","FOXO1","MYC","STAT1",'STAT3',"SMAD3"),
#                            method = "ChEA3", species = "Human")
# infer.plot = VizRegulator(infer.res = infer.res, method = "ChEA3")
}
