% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component_analysis.R
\name{ExportPCGenes}
\alias{ExportPCGenes}
\title{Export selected PC genes.}
\usage{
ExportPCGenes(pca, pc = 1:5, gene.num = 10)
}
\arguments{
\item{pca}{PCA results of \code{\link{PCA}}.}

\item{pc}{Specify PC to export genes. Default: 1:5.}

\item{gene.num}{Gene number to export for every PC. Default: 10.}
}
\value{
Dataframe contains PC, Gene and Type (Positive or Negative).
}
\description{
Export selected PC genes.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = T, sep = "\t")
meta.info <- read.table(file = meta.file, header = T)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
pca_res <- PCA(deobj = dds, transform.method = "rlog")
ExportPCGenes(pca = pca_res, pc = 1:2, gene.num = 200)
}
