% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_bind_Peak.R
\name{InteDiffQuad}
\alias{InteDiffQuad}
\title{Create Quadrant Diagram for Differential Expression Analysis of RNA-seq and Peak-related data.}
\usage{
InteDiffQuad(
  de.peak,
  peak.type = c("ChIP", "ATAC"),
  point.alpha = 0.6,
  point.size.vec = c(2, 4),
  rna.l2fc.threshold = 0,
  peak.l2fc.threshold = 0,
  linetype = 2,
  point.color.vec = c("grey", "red", "blue", "grey"),
  legend.pos = "top",
  show.corr = TRUE,
  label.num = NULL,
  label.df = NULL,
  label.color = NULL
)
}
\arguments{
\item{de.peak}{Dataframe contains integrated results of differential analysis of RNA-seq and Peak-related data.}

\item{peak.type}{Peak data type, choose from ChIP, ATAC. Default: ChIP.}

\item{point.alpha}{Opacity of a geom. Default: 0.6.}

\item{point.size.vec}{Point size for regular(DE or non-DE) and(or) labeled points.
Default: 2 for regular and 4 for labeled points.}

\item{rna.l2fc.threshold}{Log2 fold change threshold for RNA-seq to get differentially expressed results. Default: 0.}

\item{peak.l2fc.threshold}{Log2 fold change threshold for peak-related data to get differentially expressed results. Default: 0.}

\item{linetype}{Threshold linetype. Default: 2.}

\item{point.color.vec}{Point color for Down_Up, Up_Up, Down_Down, Up_Down.
Default: grey for Down_Up, red for Up_Up, blue for Down_Down and grey for Up_Down.}

\item{legend.pos}{Legend position. Default: top.}

\item{show.corr}{Logical value, whether to add pearson correlation coefficient and its significance. Default: TRUE.}

\item{label.num}{Gene number to label, according to RNA_log2FoldChange and Peak_log2FoldChange. When \code{label.df} is set NULL,
use this to determine genes to label. Default: NULL.}

\item{label.df}{Label data frame, at least contains Gene column. Default: NULL(use \code{label.num}).
When provided, the second column should not be in \code{de.peak}.}

\item{label.color}{Color for labels. Default: NULL (black).}
}
\value{
A ggplot2 object.
}
\description{
Create Quadrant Diagram for Differential Expression Analysis of RNA-seq and Peak-related data.
}
\examples{
# # prepare integrated differential expression results of RNA-seq and ATAC-seq
# # label dataframe
# label.df = data.frame(Gene = c("Ccl3", "Ccl5", "Cd28", "Cx3cr1", "Prdm1", "Tcf7", "Slamf6", "Id3", "Cxcr5"))
# # create plot
# quad.plot = InteDiffQuad(de.peak = debatac.res, peak.type = "ATAC", label.df = label.df, label.color = "green")
}
