% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NormalizedCount}
\alias{NormalizedCount}
\title{Perform Counts Normalization.}
\usage{
NormalizedCount(
  deobj,
  gtf.file = NULL,
  norm.type = c("DESeq2", "TMM", "CPM", "RPKM", "TPM")
)
}
\arguments{
\item{deobj}{Object created by DESeq2 or edgeR.}

\item{gtf.file}{Gene annotation file used to get gene length, used if \code{norm.type=="RPKM"} or \code{norm.type=="TPM"}. Default: NULL.}

\item{norm.type}{Normalization method, chosen from DESeq2, TMM, CPM, RPKM, TPM. Default: DESeq2.}
}
\value{
Matrix contains normalized counts.
}
\description{
Perform Counts Normalization.
}
\examples{
library(DESeq2)
library(DEbChIP)
count.file <- system.file("extdata", "snon_count.txt", package = "DEbChIP")
meta.file <- system.file("extdata", "snon_meta.txt", package = "DEbChIP")
count.matrix <- read.table(file = count.file, header = TRUE, sep = "\t")
meta.info <- read.table(file = meta.file, header = TRUE)
dds <- DESeq2::DESeqDataSetFromMatrix(countData = count.matrix, colData = meta.info, design = ~condition)
keep.genes <- rowSums(DESeq2::counts(dds, normalized = FALSE)) >= 10
dds <- dds[keep.genes, ]
dds$condition <- relevel(dds$condition, ref = "WT")
dds <- DESeq(dds)
NormalizedCount(dds, norm.type = "CPM")
}
